fix: ESLint config and code fixes for strict type checking
- Add .vscode-test/** to ESLint ignores (was causing 14k+ false positives)
- Configure restrict-template-expressions to allow numbers/booleans/nullish
- Set no-non-null-assertion and no-deprecated to warnings
- Disable class-literal-property-style (common config pattern)
- Allow underscore-prefixed unused variables

Code fixes:
- Remove redundant String() calls in forensic-capture.ts
- Fix exhaustive switch and nullish coalescing in agent-tree.ts
- Fix unnecessary conditions in status-bar.ts
- Remove async from estimateTotalInputTokens (no await needed)
- Fix EventEmitter.fire() calls to pass undefined for full refresh
- Fix || and ?? operator precedence with parentheses

+++ apps/vscode-ai-gateway/src/tokens/conversation-state.ts
@@ -197,7 +197,7 @@ export class ConversationStateTracker {
         parts.push({ type: "text", value: part.value });
       } else if ("data" in part && "mimeType" in part) {
         // DataPart - hash the data instead of including it
-        const data = part.data as Uint8Array;
+        const data = part.data;
         const dataHash = crypto.createHash("sha256").update(data).digest("hex");
         parts.push({
           type: "data",
